+++ src/rwlvariable.c
@@ -11,6 +11,7 @@
  *
  * History
  *
+ * bengsig  19-sep-2022 - Future keywords
  * bengsig  16-may-2022 - Debug bit 0x40 to printvar internal
  * bengsig  16-may-2022 - Flush local sql upon exit
  * bengsig  31-mar-2022 - Warn if using future sql keyword as identifier
@@ -56,6 +57,10 @@ sb4 rwladdvar2(rwl_main *rwm, text *varn, rwl_type vart, ub2 flags, text *pname)
     rwlsevere(rwm,"[rwladdvarprivate:%s;%d;0x%x]", varn, vart, flags);
     return RWL_VAR_NOGUESS;
   }
+
+  if (bit(rwm->m3flags, RWL_P3_FUTNOTIDENT))
+    rwlerror(rwm, RWL_ERROR_RESERVED_FOR_FUTURE2, varn);
+  bic(rwm->m3flags, RWL_P3_FUTNOTIDENT);
 
   switch(vart)
   {